+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CIbTZmj2.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-BoZSxRL2.js";import{_ as a}from"./mermaid.core-D-SlBfrT.js";import"./chunk-5VM5RSS4-myMQkxET.js";import"./chunk-XXDRQBXY-l3A0631u.js";import"./chunk-POPQ4Y6H-DV8TAEwP.js";import"./chunk-F27PBJKO-B7loOar0.js";import"./channel-8pnc_x5c.js";import"./index-BzevnRYd.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};